# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS API Documentation",
        "description": "Swagger API documentation for the HRMS Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://lawngreen-salamander-508885.hostingersite.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/admin/offboarding/{offboardingId}/checklists": {
            "get": {
                "tags": [
                    "Offboarding Checklist"
                ],
                "summary": "Get all checklist items for an offboarding record",
                "operationId": "getOffboardingChecklists",
                "parameters": [
                    {
                        "name": "offboardingId",
                        "in": "path",
                        "description": "Offboarding ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "offboarding_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "category_id": {
                                                "type": "string",
                                                "example": "visa_cancellation"
                                            },
                                            "task_name": {
                                                "type": "string",
                                                "example": "Submit visa cancellation to GDRFA"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "responsible_role": {
                                                "type": "string",
                                                "example": "PRO"
                                            },
                                            "notes": {
                                                "type": "string",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offboarding record not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Offboarding Checklist"
                ],
                "summary": "Create a new checklist item",
                "operationId": "createChecklist",
                "parameters": [
                    {
                        "name": "offboardingId",
                        "in": "path",
                        "description": "Offboarding ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "task_name"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "example": "general"
                                    },
                                    "task_name": {
                                        "type": "string",
                                        "example": "Return access card"
                                    },
                                    "responsible_role": {
                                        "type": "string",
                                        "example": "HR",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Collect from security desk",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checklist item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Checklist created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "offboarding_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "category_id": {
                                                    "type": "string",
                                                    "example": "general"
                                                },
                                                "task_name": {
                                                    "type": "string",
                                                    "example": "Return access card"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "responsible_role": {
                                                    "type": "string",
                                                    "example": "HR"
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/admin/checklists/{id}": {
            "put": {
                "tags": [
                    "Offboarding Checklist"
                ],
                "summary": "Update a checklist item",
                "operationId": "updateChecklist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "example": "general"
                                    },
                                    "task_name": {
                                        "type": "string",
                                        "example": "Return laptop"
                                    },
                                    "responsible_role": {
                                        "type": "string",
                                        "example": "IT",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Wipe data before returning",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checklist item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Checklist updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "offboarding_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "category_id": {
                                                    "type": "string",
                                                    "example": "general"
                                                },
                                                "task_name": {
                                                    "type": "string",
                                                    "example": "Return laptop"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "responsible_role": {
                                                    "type": "string",
                                                    "example": "IT"
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "Wipe data before returning",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Offboarding Checklist"
                ],
                "summary": "Delete a checklist item",
                "operationId": "deleteChecklist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Checklist deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}